pyproject.toml
[project]
name = "livekit-local-voice-demo"
version = "0.1.0"
description = "LiveKit Agents demo with OpenAI STT/TTS, Runpod-hosted LLM, and LiveKit SIP infrastructure."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
 "livekit-agents[openai,silero,turn-detector]~=1.4",
 "httpx>=0.28.0,<1.0.0",
 "python-dotenv>=1.0.1,<2.0.0",
]

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"
